# Core: Check parent_region.can_reach first in Location.can_reach